implement `DELETE` endpoint

+++ src/routes/Planets.ts
@@ -6,6 +6,7 @@ import { Repository } from 'typeorm';
 import HTTPStatus from 'http-status-codes';
 import { InvalidRequestBody, ResourceAlreadyExists, ResourceNotFound } from '../infra/errors';
 import { StellarSystem } from '../infra/entities/StellarSystems';
+import { throwDeprecation } from 'process';
 
 const router = express.Router();
 
@@ -145,20 +146,23 @@ router.delete("/planets/:id", async (req: Request, res: Response) => {
 
     try {
         const id: number = parseInt(req.params.id);
-        const planet: Planet = await repository.findOneByOrFail({ id: id });
+
+        const planet: Planet = await repository.findOneBy({ id: id });
+        if (!planet) {
+            throw new ResourceNotFound("Planet com ID fornecido não existe.");
+        }
+
         await repository.delete({ id: id })
 
         res.status(HTTPStatus.OK).send({
             "status": "success",
-            "message": "Requisição processada com sucesso",
-            "data": {
-                planet
-            }
+            "message": "Planet especificado deletado com sucesso.",
+            "data": {}
         })
-    } catch {
-        res.status(HTTPStatus.INTERNAL_SERVER_ERROR).send({
+    } catch (error) {
+        res.status(error.HTTPStatusCode || HTTPStatus.INTERNAL_SERVER_ERROR).send({
             "status": "error",
-            "message": "Erro ao processar a requisição",
+            "message": error.message || "Erro ao processar a requisição",
             "data": {}
         })
     }
